backend/pyproject.toml
[project]
name = "conversational-bridge-backend"
version = "0.1.0"
description = "Backend for Conversational Bridge AI"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.78.1",
    "pandas>=2.0.3",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/conversational_bridge_backend"] 
